Change variable

+++ src/HomeworkOne/Number1.java
@@ -6,11 +6,11 @@ public class Number1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Type A number: ");
-        double A = scanner.nextInt();
+        int A = scanner.nextInt();
         System.out.println("Type B number: ");
-        double B = scanner.nextInt();
+        int B = scanner.nextInt();
         System.out.println("Type C number: ");
-        double C = scanner.nextInt();
+        int C = scanner.nextInt();
         if (A > C || B > C) {
             System.out.println("Number " + C + " is between " + A + " and " + B);
         } else {
